# src/BulletInBoardServer.Controllers/BulletInBoardServer.Controllers.AnnouncementsController/src/BulletInBoardServer.Controllers.AnnouncementsController/wwwroot/openapi-original.json
{
  "openapi" : "3.0.3",
  "info" : {
    "description" : "API шлюза для управления объявлениями",
    "title" : "API Шлюз. Объявления",
    "version" : "0.0.2"
  },
  "servers" : [ {
    "url" : "/api"
  } ],
...
  "paths" : {
    "/announcements/create" : {
      "post" : {
        "operationId" : "createAnnouncement",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/createAnnouncementDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/createAnnouncementCreated"
                }
              }
            },
            "description" : "Created"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/createAnnouncement_400_response"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/createAnnouncementForbidden"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/createAnnouncementNotFound"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/createAnnouncementConflict"
                }
              }
            },
            "description" : "Conflict"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Создать объявление",
        "tags" : [ "Announcements" ]
      }
    },
    "/announcements/get-details" : {
      "get" : {
        "operationId" : "getAnnouncementDetails",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/announcementIdentifierDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/getAnnouncementDetailsOk"
                }
              }
            },
            "description" : "Ok"
          },
          "400" : {
            "description" : "Bad Request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/getAnnouncementDetailsForbidden"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/getAnnouncementDetailsNotFound"
                }
              }
            },
            "description" : "Not Found"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Получить подробности о выбранном объявлении",
        "tags" : [ "Announcements" ]
      }
    },
    "/announcements/update" : {
      "put" : {
        "operationId" : "updateAnnouncement",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/updateAnnouncementDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Ok"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/updateAnnouncementBadRequest"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/updateAnnouncementForbidden"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/updateAnnouncementNotFound"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/updateAnnouncementConflict"
                }
              }
            },
            "description" : "Conflict"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Редактировать объявление",
        "tags" : [ "Announcements" ]
      }
    },
    "/announcements/delete" : {
      "delete" : {
        "operationId" : "deleteAnnouncement",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/announcementIdentifierDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Ok"
          },
          "400" : {
            "description" : "Bad Request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/deleteAnnouncementForbidden"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/deleteAnnouncementNotFound"
                }
              }
            },
            "description" : "Not Found"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Удалить объявление",
        "tags" : [ "Announcements" ]
      }
    },
    "/announcements/published/get-list" : {
      "get" : {
        "operationId" : "getPostedAnnouncementList",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/getPostedAnnouncementListOk"
                }
              }
            },
            "description" : "Ok"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/getPostedAnnouncementListForbidden"
                }
              }
            },
            "description" : "Forbidden"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Получить список опубликованных объявлений",
        "tags" : [ "Announcements" ]
      }
    },
    "/announcements/published/hide" : {
      "post" : {
        "operationId" : "hidePostedAnnouncement",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/announcementIdentifierDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Ok"
          },
          "400" : {
            "description" : "Bad Request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/hidePostedAnnouncementForbidden"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/hidePostedAnnouncementNotFound"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/hidePostedAnnouncementConflict"
                }
              }
            },
            "description" : "Conflict"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Скрыть опубликованное объявление",
        "tags" : [ "Announcements" ]
      }
    },
    "/announcements/hidden/get-list" : {
      "get" : {
        "operationId" : "getHiddenAnnouncementList",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/getHiddenAnnouncementListOk"
                }
              }
            },
            "description" : "Ok"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/getHiddenAnnouncementListForbidden"
                }
              }
            },
            "description" : "Forbidden"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Получить список скрытых объявлений",
        "tags" : [ "Announcements" ]
      }
    },
    "/announcements/hidden/restore" : {
      "post" : {
        "operationId" : "restoreHiddenAnnouncement",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/announcementIdentifierDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Ok"
          },
          "400" : {
            "description" : "Bad Request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/restoreHiddenAnnouncementForbidden"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/restoreHiddenAnnouncementNotFound"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/restoreHiddenAnnouncementConflict"
                }
              }
            },
            "description" : "Conflict"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Восстановить скрытое объявление",
        "tags" : [ "Announcements" ]
      }
    },
    "/announcements/delayed-publishing/get-list" : {
      "get" : {
        "operationId" : "getDelayedPublishingAnnouncementList",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/getDelayedPublishingAnnouncementListOk"
                }
              }
            },
            "description" : "Ok"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/getDelayedPublishingAnnouncementListForbidden"
                }
              }
            },
            "description" : "Forbidden"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Получить список объявлений, ожидающих отложенную публикацию",
        "tags" : [ "Announcements" ]
      }
    },
    "/announcements/delayed-publishing/publish-immediately" : {
      "post" : {
        "operationId" : "publishImmediatelyDelayedPublishingAnnouncement",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/announcementIdentifierDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Ok"
          },
          "400" : {
            "description" : "Bad Request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/publishImmediatelyDelayedPublishingAnnouncementForbidden"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/publishImmediatelyDelayedPublishingAnnouncementNotFound"
                }
              }
            },
            "description" : "Not Found"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Сразу опубликовать отложенное объявление",
        "tags" : [ "Announcements" ]
      }
    },
    "/announcements/delayed-hidden/get-list" : {
      "get" : {
        "operationId" : "getDelayedHiddenAnnouncementList",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/getDelayedHiddenAnnouncementListOk"
                }
              }
            },
            "description" : "Ok"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/getDelayedHiddenAnnouncementListForbidden"
                }
              }
            },
            "description" : "Forbidden"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Получить список объявлений, ожидающих отложенное сокрытие",
        "tags" : [ "Announcements" ]
      }
    }
  },
  "components" : {
    "schemas" : {
      "createAnnouncementDto" : {
        "properties" : {
          "content" : {
            "description" : "Текстовое содержимое объявления",
            "type" : "string"
          },
          "userIds" : {
            "description" : "Идентификаторы пользователей, для которых создается объявление",
            "items" : {
              "format" : "uuid",
              "type" : "string"
            },
            "minItems" : 1,
            "type" : "array"
          },
          "categoryIds" : {
            "description" : "Идентификаторы категорий объявлений",
            "items" : {
              "format" : "uuid",
              "type" : "string"
            },
            "type" : "array"
          },
          "attachmentIds" : {
            "description" : "Идентификаторы вложений, прикрепляемых к объявлению",
            "items" : {
              "format" : "uuid",
              "type" : "string"
            },
            "minItems" : 1,
            "nullable" : true,
            "type" : "array"
          },
          "delayedPublishingAt" : {
            "description" : "Срок отложенной публикации объявления",
            "format" : "date-time",
            "nullable" : true,
            "type" : "string"
          },
          "delayedHidingAt" : {
            "description" : "Срок отложенного сокрытия объявления",
            "format" : "date-time",
            "nullable" : true,
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "announcementIdentifierDto" : {
        "description" : "Идентификатор объявления",
        "format" : "uuid",
        "type" : "string"
      },
...
      "createAnnouncementCreated" : {
        "example" : {
          "content" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
        },
        "properties" : {
          "content" : {
            "description" : "Идентификатор объявления",
            "format" : "uuid",
            "type" : "string"
          }
        },
        "title" : "createAnnouncementCreated",
        "type" : "object"
      },
      "createAnnouncement_400_response" : {
        "properties" : {
          "code" : {
            "$ref" : "#/components/schemas/createAnnouncementResponses"
          }
        },
        "type" : "object"
      },
      "createAnnouncementForbidden" : {
        "properties" : {
          "code" : {
            "$ref" : "#/components/schemas/createAnnouncementResponses"
          }
        },
        "title" : "createAnnouncementForbidden",
        "type" : "object"
      },
      "createAnnouncementNotFound" : {
        "properties" : {
          "code" : {
            "$ref" : "#/components/schemas/createAnnouncementResponses"
          }
        },
        "title" : "createAnnouncementNotFound",
        "type" : "object"
      },
      "createAnnouncementConflict" : {
        "properties" : {
          "code" : {
            "$ref" : "#/components/schemas/createAnnouncementResponses"
          }
        },
        "title" : "createAnnouncementConflict",
        "type" : "object"
      },
      "getAnnouncementDetailsOk" : {
        "example" : {
          "content" : {
            "hiddenAt" : "2000-01-23T04:56:07.000+00:00",
            "audienceSize" : 0,
            "publishedAt" : "2000-01-23T04:56:07.000+00:00",
            "authorName" : "authorName",
            "viewsCount" : 0,
            "surveys" : [ {
              "isAnonymous" : false,
              "isOpen" : true,
              "questions" : [ {
                "votersAmount" : 0,
                "answers" : [ {
                  "votersAmount" : 0,
                  "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
                  "content" : "content"
                }, {
                  "votersAmount" : 0,
                  "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
                  "content" : "content"
                } ],
                "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
                "isMultipleChoiceAllowed" : false,
                "content" : "content"
              }, {
                "votersAmount" : 0,
                "answers" : [ {
                  "votersAmount" : 0,
                  "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
                  "content" : "content"
                }, {
                  "votersAmount" : 0,
                  "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
                  "content" : "content"
                } ],
                "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
                "isMultipleChoiceAllowed" : false,
                "content" : "content"
              } ],
              "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
              "voteFinishedAt" : "2000-01-23T04:56:07.000+00:00"
            }, {
              "isAnonymous" : false,
              "isOpen" : true,
              "questions" : [ {
                "votersAmount" : 0,
                "answers" : [ {
                  "votersAmount" : 0,
                  "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
                  "content" : "content"
                }, {
                  "votersAmount" : 0,
                  "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
                  "content" : "content"
                } ],
                "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
                "isMultipleChoiceAllowed" : false,
                "content" : "content"
              }, {
                "votersAmount" : 0,
                "answers" : [ {
                  "votersAmount" : 0,
                  "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
                  "content" : "content"
                }, {
                  "votersAmount" : 0,
                  "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
                  "content" : "content"
                } ],
                "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
                "isMultipleChoiceAllowed" : false,
                "content" : "content"
              } ],
              "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
              "voteFinishedAt" : "2000-01-23T04:56:07.000+00:00"
            } ],
            "delayedHidingAt" : "2000-01-23T04:56:07.000+00:00",
            "delayedPublishingAt" : "2000-01-23T04:56:07.000+00:00",
            "files" : [ {
              "name" : "name",
              "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
              "type" : "mediafile"
            }, {
              "name" : "name",
              "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
              "type" : "mediafile"
            } ],
            "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
            "categories" : [ {
              "color" : "color",
              "name" : "name",
              "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
            }, {
              "color" : "color",
              "name" : "name",
              "id" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
            } ],
            "content" : "content"
          }
        },
        "properties" : {
          "content" : {
            "$ref" : "#/components/schemas/announcementDetailsDto"
          }
        },
        "title" : "getAnnouncementDetailsOk",
        "type" : "object"
      },
      "getAnnouncementDetailsForbidden" : {
        "properties" : {
          "code" : {
            "$ref" : "#/components/schemas/getAnnouncementDetailsResponses"
          }
        },
        "title" : "getAnnouncementDetailsForbidden",
        "type" : "object"
      },
...
    }
  }
}